PCLParaphernalia will no longer throw an exception if a drive disappears (i. e. an removable USB drive).
PCLParaphernalia stores the path of the last opened file in the Windows registry. If the drive letter goes away and the user tries to open a file (i. e. in the PRN File Analysis Tool), PCLParaphernalia has thrown an exception. This is fixed. Now if the directory of the last opened file does not exist (note that the driver letter is part of the diretory path), PCLParaphernalia uses the users "My Documents" folder instead.