Allow conversion of licensed fonts.
The "os2" table of a TrueType font specifies if conversion of a licensed font is allowed or not (see field "fsType"). If a special bit in this field is set, PCL Paraphernalia now asks the user if the user got permission from the legal owner. Conversion only continues if the user confirms that he/she got the permission.